refactor: replace icon/icon keys import, remove hack on StyledBridgeChain for icon
