+++ src/common/commonTypes.ts
@@ -1,13 +1,9 @@
 import { ToastPosition, ToastTransition } from "react-toastify/dist/types";
 
-import icons from "./components/Icon";
-
 export interface IStyleableProps {
   style?: any;
   className?: string;
 }
-
-export type IconKeys = keyof typeof icons;
 
 export type NotificationType = {
   position: ToastPosition;
